requirements.txt
# Core
pytz==2020.5
Django==3.1.5
gunicorn==20.0.4
django-environ==0.4.5

# For the persistence stores
psycopg2-binary==2.8.6

# Rest apis
djangorestframework==3.12.2
Pillow==8.1.0
pyjwt==2.0.1

# Developer Tools
flake8==3.8.4
